+++ shepherd_core/shepherd_core/commons.py
@@ -1,8 +1,6 @@
 """Container for commonly shared constants."""
 
-SAMPLERATE_SPS_DEFAULT: int = 100_000
+from .config import config
 
-UID_NAME: str = "SHEPHERD_NODE_ID"
-UID_SIZE: int = 2
-
-TESTBED_SERVER_URI: str = "http://127.0.0.1:8000/shepherd"
+# TODO: deprecated - replace with config in subprojects and then remove here
+SAMPLERATE_SPS_DEFAULT: int = config.SAMPLERATE_SPS